fix(warn): Fix more C4018 from debug builds.

+++ Code/Commando/NAT.cpp
@@ -3064,7 +3064,7 @@ int FirewallHelperClass::Negotiate_Port(void)
 				if (resend_timer < TIMEGETTIME()) {
 					//if (!CrapFlag)
 					const size_t payload_size = ::strlen(packet) + 1;
-					WWASSERT(payload_size <= std::numeric_limits<int>::max());
+					WWASSERT(payload_size <= size_t(std::numeric_limits<int>::max()));
 					WOLNATInterface.Send_Game_Format_Packet_To(&PlayersFirewallAddress, packet, static_cast<int>(payload_size), socket);
 					resend_timer = TIMEGETTIME() + (TIMER_SECOND / 2);
 				}
